FInished the pieces dictionary. Next is the moving piece logic.

+++ TetrisCSharp/GameStatus/Position.cs
@@ -3,10 +3,10 @@ namespace TetrisCSharp.GameStatus
 {
     public struct Position
     {
-        public byte row {get; }
-        public byte column {get; }
+        public short row {get; }
+        public short column {get; }
 
-        public Position(byte row, byte column)
+        public Position(short row, short column)
         {
             this.row = row;
             this.column = column;
